update alliance to version 5.1.1 to fix compilation errors on macOS 10.11+
update alliance to version 5.1.1 to fix compilation errors on macOS Yosemite, Sierra and High Sierra. The old Alliance 5.0 does not compile on these systems, As Newer Xcode assumes C++11 by default, while Alliance 5.0 uses the non-standard header <hash_map> which name was changed to <unordered_map> when it was standardized in C++11. Alliance 5.1.1 fixes all this and more.